LayoutTests/imported/w3c/web-platform-tests/resource-timing/resource_TAO_cross_origin_redirect_chain-expected.txt

Blocked access to external URL http://www.localhost:8800/resource-timing/resources/multi_redirect.py?page_origin=http://localhost:8800&cross_origin=http://www.localhost:8800&timing_allow=1
CONSOLE MESSAGE: line 26: TypeError: undefined is not an object (evaluating 'entry.redirectStart')


Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'entry.redirectStart')

PASS window.performance is defined 
PASS window.performance.getEntriesByName is defined 
FAIL There should be one entry. assert_equals: There should be one entry. expected 1 but got 0
